improve printing the window output. improved displayable class

+++ src/graphics/ISectionComponent.cpp
@@ -1,7 +1,6 @@
 #include "headers/ISectionComponent.h"
 
 ISectionComponent::ISectionComponent(size_t size) : m_size(size), m_section_drawing(size) {}
-
 ISectionComponent::~ISectionComponent() {}
 
 void ISectionComponent::set_parent_pane(Pane* parent_pane) {
@@ -11,7 +10,7 @@ void ISectionComponent::set_parent_pane(Pane* parent_pane) {
   this->parent_pane = parent_pane;
 }
 
-int ISectionComponent::size() {
+size_t ISectionComponent::size() {
   return m_size;
 }
