+++ src/graphics/headers/section.h
@@ -15,13 +15,13 @@ class Section : public ISectionComponent {
   explicit Section(string name, size_t size);
   virtual ~Section();
   const string* at(int index);
-  void set_content_at_index(string, const int&);
+  void set_content_at_index(string, const size_t&);
   void set_content(std::initializer_list<string>);
   void set_content(const vector<string>*);
   string get_name();
 
  private:
-  bool is_valid_index(const int& index);
+  bool is_valid_index(const size_t& index);
   string m_name;
 };
