+++ src/graphics/section.cpp
@@ -10,14 +10,14 @@ Section::Section(string name, size_t size) :
 
 Section::~Section() {}
 
-void Section::set_content_at_index(string content, const int& index) {
+void Section::set_content_at_index(string content, const size_t& index) {
   if (!is_valid_index(index))
     return;
 
   m_section_drawing[index] = content;
 }
 
-bool Section::is_valid_index(const int& index) {
+bool Section::is_valid_index(const size_t& index) {
   if (index >= size()) return false;
   if (index < 0) return false;
   return true;
